Moving macro to `pallas-codec`. transparent index no longer necessary
